pyproject.toml
[project]
name = "google-map-crawler"
version = "0.1.0"
description = ""
authors = [
    {name = "Boso",email = "31940520+linboso@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "selenium (>=4.27.1,<5.0.0)",
    "regex (>=2024.11.6,<2025.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "deep-translator (>=1.11.4,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "geopy (>=2.4.1,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "geopandas (>=1.0.1,<2.0.0)",
    "shapely (>=2.0.6,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
